refactor: move Solana scripts to separate private package
Moves governance scripts from solana/ts/scripts/ into solana/ts-scripts/
with its own package.json (private: true) to avoid polluting the
published @wormhole-foundation/sdk-solana-ntt SDK with script-only
dependencies like @xlabs-xyz/ledger-signer-solana and dotenv.

Mirrors the existing evm/ts-scripts/ pattern.
