+++ solana/package.json
@@ -44,8 +44,6 @@
   },
   "devDependencies": {
     "@types/bn.js": "^5.1.5",
-    "@xlabs-xyz/ledger-signer-solana": "^0.0.1",
-    "dotenv": "^16.4.5",
     "nock": "^13.3.3",
     "ts-jest": "^29.1.2",
     "ts-node": "^10.9.2",
