+++ solana/ts-scripts/config/devnet/programs.json
