+++ solana/ts-scripts/config/mainnet/governance-vaas-june-05.json
