+++ solana/ts-scripts/config/mainnet/programs.json
