+++ solana/ts-scripts/src/helpers.ts
